+++ R/fit.R
@@ -505,7 +505,8 @@ add_predictions <- function(fit, grid = NULL, time = NULL) {
 
   ## Make preference functions --------------------------
   pref_funcs <- .make_pref_funcs(par_est$alpha, par_est$beta, par_est$gamma,
-                                 dat$knots_tax, dat$knots_dif)
+                                 dat$knots_tax, dat$knots_dif,
+                                 method = conf$smooth_method)
 
 
   ## Local interpolation --------------------------------